+++ app/src/main/res/xml/file_paths.xml
@@ -1,4 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
+<!--
+SPDX-FileCopyrightText: NOI Techpark <digital@noi.bz.it>
+
+SPDX-License-Identifier: CC0-1.0
+-->
 <paths>
 	<cache-path name="shared_vcard" path="/" />
 </paths>
